Extended Bootstrap Framework with Sprite and Text
Added Sprite and Textbox objects which inherit from Component and can be placed using the grid-based framework I made in Screen.cs (and Row.cs). The Sprite class is any image we want to add (that doesn't need click functionality like a button). The Textbox class is for adding text to the screen based on the grid system (so pixel math doesn't have to be used).